backend/requirements.txt
# Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
